+++ frontend/src/pages/Clients/ClientsPage.js
@@ -342,6 +342,24 @@ class ClientsPage extends Component {
                                     )}
                                 </div>
                             ))}
+                            {/* Ajout de la pagination ici */}
+                            <div className={stylesMobile.pagination}>
+                                <button
+                                    onClick={this.handlePreviousPage}
+                                    disabled={currentPage === 1}
+                                >
+                                    <i className="fa-solid fa-chevron-left"></i>
+                                </button>
+                                <span>
+                                    Page {currentPage} / {totalPages}
+                                </span>
+                                <button
+                                    onClick={this.handleNextPage}
+                                    disabled={currentPage === totalPages}
+                                >
+                                    <i className="fa-solid fa-chevron-right"></i>
+                                </button>
+                            </div>
                         </div>
                     </>
                 ) : (